refactor: 매직 넘버 상수화
