+++ src/main/java/roomescape/controller/rest/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.rest.normal;
+package roomescape.controller.rest;
 
 import jakarta.validation.Valid;
 import java.time.LocalDate;
@@ -13,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.service.ReservationService;
 import roomescape.service.request.ReservationCreateRequest;
 import roomescape.service.response.ReservationResponse;
-import roomescape.service.ReservationService;
 
 @RestController
 @RequestMapping("/reservations")
