+++ src/main/java/roomescape/domain/member/MemberEmail.java
@@ -8,6 +8,7 @@ public record MemberEmail(String email) {
     private static final Pattern EMAIL_PATTERN = Pattern.compile(
             "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"
     );
+    public static final int MAXIMUM_EMAIL_LENGTH = 40;
 
     public MemberEmail(final String email) {
         this.email = Objects.requireNonNull(email, "email은 null이 아니어야 합니다.");
@@ -20,7 +21,7 @@ public record MemberEmail(String email) {
             throw new IllegalStateException("올바른 이메일 형식이 아닙니다: " + email);
         }
 
-        if (email.length() > 40) {
+        if (email.length() > MAXIMUM_EMAIL_LENGTH) {
             throw new IllegalStateException("email은 40자 이하여야 합니다.");
         }
     }
